A hard day's work
this was so exhausting. It took me the entire day, and now I'm so tired I can barely move. Not only did I have to press the ctrl key, but also c at the same time. Then the same with v. I'm going to burn out for sure after this